Update helper.py
Fix test connection failing for local Qdrant.

By default SimpleDirectoryReader discards paths which contain one or more parts that are hidden. In local, packages could be installed in a venv. This means a path can contain a .venv in it which will then be treated as hidden and subsequently discarded.

Signed-off-by: Rahul Johny <116638720+johnyrahul@users.noreply.github.com>